+++ tests/Sample.CounterChainExecutor.UnitTests/Usings.cs
@@ -1,0 +1,4 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Neo;
+global using Neo.L2;
+global using Sample.CounterChainExecutor;